moved vars to environment.js refactored actions
