+++ src/store/actions/actionTypes.js
@@ -1,5 +1,4 @@
 export const FETCH_SEARCH_RESULT = 'FETCH_SEARCH_RESULT';
-export const FETCH_SEARCH_RESULT_SUCCESS = 'FETCH_SEARCH_RESULT_SUCCESS';
 export const FETCH_SEARCH_RESULT_FAILURE = 'FETCH_SEARCH_RESULT_FAILURE';
 
 export const FETCH_PEOPLE = 'FETCH_PEOPLE';
@@ -33,5 +32,3 @@ export const STOP_SAVING_RECENT_SEARCHES = 'STOP_SAVING_RECENT_SEARCHES';
 export const MODAL_VISIBLE = 'MODAL_VISIBLE';
 export const GET_INFO = 'GET_INFO';
 export const STOP_SEARCH_ME = 'STOP_SEARCH_ME';
-
-
